# openapi.yaml
openapi: 3.0.3
info:
  title: Telegram Post Generator API
  version: 1.0.0
  description: >
    REST API для управления пользователями, их настройками, постами и автоматической публикацией в Telegram-каналы.
servers:
  - url: http://localhost:4000/api
    description: Локальный сервер разработки
...
paths:
...
  /posts/upload:
    post:
      tags: [Posts]
      summary: Загрузка изображения
      security:
        - AuthTokenHeader: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Успешная загрузка
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  url: { type: string }
        "400":
          description: Ошибка загрузки или отсутствует токен
  /posts:
    post:
      tags: [Posts]
      summary: Создание нового поста
      security:
        - AuthTokenHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCreate"
      responses:
        "200":
          description: Пост успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: integer }
        "403":
          description: Доступ запрещён
        "400":
          description: Ошибка данных
  /posts/user/{userId}:
    get:
      tags: [Posts]
      summary: Получение всех постов пользователя
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Список постов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "403":
          description: Доступ запрещён
  /posts/{id}:
    get:
      tags: [Posts]
      summary: Получить пост по ID
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Пост найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "403":
          description: Доступ запрещён
        "404":
          description: Пост не найден

    put:
      tags: [Posts]
      summary: Обновление поста
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Пост обновлён
        "403":
          description: Доступ запрещён
        "404":
          description: Пост не найден

    delete:
      tags: [Posts]
      summary: Удаление поста
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Пост удалён
        "403":
          description: Доступ запрещён
        "404":
          description: Пост не найден
  /posts/{id}/image:
    delete:
      tags: [Posts]
      summary: Удалить изображение из поста
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Изображение удалено
        "403":
          description: Доступ запрещён
        "404":
          description: Пост не найден
  /posts/schedulePost/{id}:
    post:
      tags: [Posts]
      summary: Запланировать пост
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduledAt]
              properties:
                scheduledAt:
                  type: string
                  format: date-time
                  example: "2025-10-09T12:00:00Z"
      responses:
        "200":
          description: Пост запланирован
        "403":
          description: Доступ запрещён
        "404":
          description: Пост не найден

    delete:
      tags: [Posts]
      summary: Отменить запланированную публикацию
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Запланированная публикация отменена
        "403":
          description: Доступ запрещён
        "404":
          description: Пост не найден
  /posts/sendPost/{id}:
    post:
      tags: [Posts]
      summary: Отправить пост вручную
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Пост отправлен
        "403":
          description: Доступ запрещён
        "404":
          description: Пост не найден
  /posts/generate-posts/{userId}:
    post:
      tags: [Posts]
      summary: Сгенерировать посты на основе пользовательских настроек
      security:
        - AuthTokenHeader: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Сгенерированные посты
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  userId: { type: integer }
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
        "403":
          description: Доступ запрещён
components:
  securitySchemes:
    AuthTokenHeader:
      type: apiKey
      in: header
      name: x-auth-token
      description: >
        Уникальный токен авторизации пользователя.
        Передаётся в каждом запросе, где требуется проверка владельца (в header x-auth-token).
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
          example: 12
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: "Заголовок поста"
        description:
          type: string
          example: "Описание поста для Telegram"
        url:
          type: string
          nullable: true
          example: "https://placehold.co/600x400"
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-10-09T12:00:00Z"
        sent:
          type: integer
          example: 0
    PostCreate:
      type: object
      required: [userId, title]
      properties:
        userId:
          type: integer
          example: 1
        title:
          type: string
          example: "Мой первый пост"
        description:
          type: string
          example: "Это тестовое описание"
        url:
          type: string
          nullable: true
          example: "https://placehold.co/600x400"
        scheduledAt:
          type: string
          format: date-time
          nullable: true
    PostUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        url:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true